Replace ViewModelTests. Pass CanRollValuesLessThanOrEqualToTen.
